Added combined purple and green blob detection.
Added and combined the detection for the green blobs. Still need to fine tune the color bounds.

Still need to also find the bounds within the image to search (I'm not sure if I will go for a pure rectangle, of if I will define a function to look along a rotated rectangle (more precision).