Updated a few Opmodes and Gradle
Updated a few Opmodes and Gradle. Still working on fine tuning the object detection in camera, but this is just a backup/savestate as the ground detection has been implemented into the opmode. Again, there is still a lot that needs to be modified/tweaked to get it to work.